did some changes to prepare for testing and simplifications
